# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-grade REST API for product catalog management with categories, variants, and advanced filtering",
        "title": "Go Product Catalog API",
        "contact": {
            "name": "Alex Necsoiu",
            "email": "axel.necsoiu@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/catalog": {
            "get": {
                "description": "Returns a paginated list of all products with support for category filtering and price range searching. Results include pagination metadata (total count, offset, limit).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List products with pagination and filters",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items to return, max 100 (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category code (case-insensitive, e.g., 'clothing')",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Filter products with price less than this value (e.g., 100.50)",
                        "name": "priceLessThan",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products with pagination metadata",
                        "schema": {
                            "$ref": "#/definitions/catalog.CatalogResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error: invalid offset, limit, or price parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching products",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/catalog/{code}": {
            "get": {
                "description": "Returns complete product information including all variants and category details. Variant prices inherit from the product price if not explicitly set.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Get product details by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product code (e.g., 'PROD001')",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details with variants and category information",
                        "schema": {
                            "$ref": "#/definitions/catalog.ProductDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error: missing or invalid product code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching product",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Returns a complete list of all product categories ordered by code for deterministic results. Categories define the product taxonomy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List all categories",
                "responses": {
                    "200": {
                        "description": "List of all categories",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error while fetching categories",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product category with validation and uniqueness constraints. Category codes must be alphanumeric (uppercase), hyphens, and underscores.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error: missing required fields or invalid code format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict: category code already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error while creating category",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "catalog.CatalogListData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.ProductItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "catalog.CatalogResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/catalog.CatalogListData"
                }
            }
        },
        "catalog.CategoryItem": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "catalog.ProductDetail": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/catalog.CategoryItem"
                },
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.VariantItem"
                    }
                }
            }
        },
        "catalog.ProductDetailData": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/catalog.ProductDetail"
                }
            }
        },
        "catalog.ProductDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/catalog.ProductDetailData"
                }
            }
        },
        "catalog.ProductItem": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/catalog.CategoryItem"
                },
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                }
            }
        },
        "catalog.VariantItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                }
            }
        },
        "categories.CategoriesListData": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/categories.CategoryItem"
                    }
                }
            }
        },
        "categories.CategoriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/categories.CategoriesListData"
                }
            }
        },
        "categories.CategoryItem": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "categories.CategoryRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ELECTRONICS"
                },
                "name": {
                    "type": "string",
                    "example": "Electronics \u0026 Gadgets"
                }
            }
        },
        "categories.CategoryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/categories.CategoryItem"
                }
            }
        }
    }
}